+++ src/types/chat_member.rs
@@ -1,3 +1,5 @@
+use std::ops::Deref;
+
 use serde::{Deserialize, Serialize};
 
 use crate::types::User;
@@ -44,55 +46,49 @@ pub struct Administrator {
     /// Custom title for this user.
     pub custom_title: Option<String>,
 
-    /// True, if the user's presence in the chat is hidden
+    /// `true` if the user's presence in the chat is hidden
     pub is_anonymous: bool,
 
-    /// `true`, if the bot is allowed to edit
-    /// administrator privileges of that user.
+    /// `true` if the bot is allowed to edit administrator privileges of that
+    /// user.
     pub can_be_edited: bool,
 
-    /// `true`, if the administrator can access the chat event log, chat
+    /// `true` if the administrator can access the chat event log, chat
     /// statistics, message statistics in channels, see channel members, see
     /// anonymous administrators in supergroups and ignore slow mode. Implied by
     /// any other administrator privilege
     pub can_manage_chat: bool,
 
-    /// `true`, if the administrator can change the chat
-    /// title, photo and other settings.
+    /// `true` if the administrator can change the chat title, photo and other
+    /// settings.
     pub can_change_info: bool,
 
-    /// `true`, if the administrator can post in the
-    /// channel, channels only.
+    /// `true` if the administrator can post in the channel, channels only.
     pub can_post_messages: Option<bool>,
 
-    /// `true`, if the administrator can edit messages of
-    /// other users and can pin messages, channels only.
+    /// `true` if the administrator can edit messages of other users and can pin
+    /// messages, channels only.
     pub can_edit_messages: Option<bool>,
 
-    /// `true`, if the administrator can delete messages
-    /// of other users.
+    /// `true` if the administrator can delete messages of other users.
     pub can_delete_messages: bool,
 
-    /// `true`, if the administrator can manage voice chats.
+    /// `true` if the administrator can manage voice chats.
     pub can_manage_voice_chats: bool,
 
-    /// `true`, if the administrator can invite new users
-    /// to the chat.
+    /// `true` if the administrator can invite new users to the chat.
     pub can_invite_users: bool,
 
-    /// `true`, if the administrator can restrict,
-    /// ban or unban chat members.
+    /// `true` if the administrator can restrict, ban or unban chat members.
     pub can_restrict_members: bool,
 
-    /// `true`, if the administrator can pin messages,
-    /// supergroups only.
+    /// `true` if the administrator can pin messages, supergroups only.
     pub can_pin_messages: Option<bool>,
 
-    /// `true`, if the administrator can add new
-    /// administrators with a subset of his own privileges or demote
-    /// administrators that he has promoted, directly or indirectly
-    /// (promoted by administrators that were appointed by the
-    /// user).
+    /// `true` if the administrator can add new administrators with a subset of
+    /// his own privileges or demote administrators that he has promoted,
+    /// directly or indirectly (promoted by administrators that were appointed
+    /// by the user).
     pub can_promote_members: bool,
 }
 
@@ -100,24 +96,23 @@ pub struct Administrator {
 /// enum.
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub struct Restricted {
-    /// Date when restrictions will be lifted for
-    /// this user, unix time.
-    pub until_date: i32,
+    /// Date when restrictions will be lifted for this user, unix time.
+    pub until_date: i64,
 
-    /// Restricted only. `true`, if the user can send text messages,
-    /// contacts, locations and venues.
+    /// `true` if the user can send text messages, contacts, locations and
+    /// venues.
     pub can_send_messages: bool,
 
-    /// Restricted only. `true`, if the user is allowed to send audios,
-    /// documents, photos, videos, video notes and voice notes.
+    /// `true` if the user is allowed to send audios, documents, photos, videos,
+    /// video notes and voice notes.
     pub can_send_media_messages: bool,
 
-    /// Restricted only. `true`, if the user is allowed to send animations,
-    /// games, stickers and use inline bots.
+    /// `true` if the user is allowed to send animations, games, stickers and
+    /// use inline bots.
     pub can_send_other_messages: bool,
 
-    /// Restricted only. `true`, if the user is allowed to add web page
-    /// previews to their messages.
+    /// `true` if the user is allowed to add web page previews to their
+    /// messages.
     pub can_add_web_page_previews: bool,
 }
 
@@ -125,14 +120,33 @@ pub struct Restricted {
 /// enum.
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub struct Kicked {
-    /// Date when restrictions will be lifted for
-    /// this user, unix time.
-    pub until_date: i32,
+    /// Date when restrictions will be lifted for this user, unix time.
+    pub until_date: i64,
 }
 
-impl ChatMember {
+/// This allows calling [`ChatMemberKind`]'s methods directly on [`ChatMember`].
+///
+/// ```no_run
+/// use teloxide_core::types::ChatMember;
+///
+/// let member: ChatMember = todo!();
+///
+/// let _ = member.status();
+/// let _ = member.kind.status();
+/// ```
+impl Deref for ChatMember {
+    type Target = ChatMemberKind;
+
+    fn deref(&self) -> &Self::Target {
+        &self.kind
+    }
+}
+
+/// Simple methods for checking a user status.
+impl ChatMemberKind {
+    /// Returns chat member status.
     pub fn status(&self) -> ChatMemberStatus {
-        match &self.kind {
+        match self {
             ChatMemberKind::Creator(_) => ChatMemberStatus::Creator,
             ChatMemberKind::Administrator(_) => ChatMemberStatus::Administrator,
             ChatMemberKind::Member => ChatMemberStatus::Member,
@@ -140,6 +154,74 @@ impl ChatMember {
             ChatMemberKind::Left => ChatMemberStatus::Left,
             ChatMemberKind::Kicked(_) => ChatMemberStatus::Kicked,
         }
+    }
+
+    /// Returns `true` if the user is the [creator] (owner) of the given chat.
+    ///
+    /// [creator]: ChatMemberKind::Creator
+    pub fn is_creator(&self) -> bool {
+        matches!(self, Self::Creator { .. })
+    }
+
+    /// Returns `true` if the user is an [administrator] of the given chat.
+    ///
+    /// [administrator]: ChatMemberKind::Administrator
+    ///
+    /// **Note**: this function **doesn't** return `true` if the user is the
+    /// creator of the given chat. See also: [`is_privileged`].
+    ///
+    /// [`is_privileged`]: ChatMemberKind::is_privileged
+    pub fn is_administrator(&self) -> bool {
+        matches!(self, Self::Administrator { .. })
+    }
+
+    /// Returns `true` if the user is a common [member] of the given chat.
+    ///
+    /// [member]: ChatMemberKind::Member
+    pub fn is_member(&self) -> bool {
+        matches!(self, Self::Member { .. })
+    }
+
+    /// Returns `true` if the user is [restricted] in the given chat.
+    ///
+    /// [restricted]: ChatMemberKind::Restricted
+    pub fn is_restricted(&self) -> bool {
+        matches!(self, Self::Restricted { .. })
+    }
+
+    /// Returns `true` if the user [left] the given chat.
+    ///
+    /// [left]: ChatMemberKind::Left
+    pub fn is_left(&self) -> bool {
+        matches!(self, Self::Left { .. })
+    }
+
+    /// Returns `true` if the user is [kicked] from the given chat.
+    ///
+    /// [kicked]: ChatMemberKind::Kicked
+    pub fn is_kicked(&self) -> bool {
+        matches!(self, Self::Kicked { .. })
+    }
+}
+
+/// Compound methods for checking a user status.
+impl ChatMemberKind {
+    /// Returns `true` if the user is privileged in the given chat. i.e. if the
+    /// user is either the [creator] or an [administrator] in the given chat.
+    ///
+    /// [creator]: ChatMemberKind::Creator
+    /// [administrator]: ChatMemberKind::Administrator
+    pub fn is_privileged(&self) -> bool {
+        self.is_administrator() || self.is_creator()
+    }
+
+    /// Returns `true` if the user is currently present in the chat. i.e. if the
+    /// user **hasn't** [left] or been [kicked].
+    ///
+    /// [left]: ChatMemberKind::Left
+    /// [kicked]: ChatMemberKind::Kicked
+    pub fn is_present(&self) -> bool {
+        !(self.is_left() || self.is_kicked())
     }
 }
 
@@ -154,227 +236,328 @@ impl ChatMemberKind {
         }
     }
 
+    /// Returns `true` if the user's presence in the chat is hidden.
+    ///
+    /// I.e. returns `true` if the user is the creator of the chat or an
+    /// administrator in the chat and has the [`can_manage_chat`] privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_manage_chat`]: Administrator::can_manage_chat
+    pub fn is_anonymous(&self) -> bool {
+        match self {
+            Self::Creator(Creator { is_anonymous, .. })
+            | Self::Administrator(Administrator { is_anonymous, .. }) => *is_anonymous,
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => false,
+        }
+    }
+
     /// Getter for [`Restricted::until_date`] and [`Kicked::until_date`] fields.
-    pub fn until_date(&self) -> Option<i32> {
+    pub fn until_date(&self) -> Option<i64> {
         match &self {
             Self::Creator(_) | Self::Administrator(_) | Self::Member | Self::Left => None,
             Self::Restricted(Restricted { until_date, .. })
             | Self::Kicked(Kicked { until_date, .. }) => Some(*until_date),
         }
     }
+}
 
-    /// Getter for [`Administrator::can_be_edited`] field.
-    pub fn can_be_edited(&self) -> Option<bool> {
-        match &self {
-            Self::Administrator(Administrator { can_be_edited, .. }) => Some(*can_be_edited),
+/// Methods for checking admin privileges.
+impl ChatMemberKind {
+    /// Returns `true` if the user is an administrator in the given chat and the
+    /// bot is allowed to edit administrator privileges of that user.
+    pub fn can_be_edited(&self) -> bool {
+        match self {
+            Self::Administrator(Administrator { can_be_edited, .. }) => *can_be_edited,
+            // Creator can't ever be edited by any bot.
             Self::Creator(_)
             | Self::Member
             | Self::Restricted(_)
             | Self::Left
-            | Self::Kicked(_) => None,
+            | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Administrator::can_manage_chat`] field.
-    pub fn can_manage_chat(&self) -> Option<bool> {
-        match &self {
+    /// Returns `true` if the user can access the chat event log, chat
+    /// statistics, message statistics in channels, see channel members, see
+    /// anonymous administrators in supergroups and ignore slow mode. Implied by
+    /// any other administrator privilege.
+    ///
+    /// I.e. returns `true` if the user
+    /// - is the creator of the chat
+    /// - is an administrator in the given chat and has [`can_manage_chat`]
+    ///   privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_manage_chat`]: Administrator::can_manage_chat
+    pub fn can_manage_chat(&self) -> bool {
+        match self {
+            Self::Creator(_) => true,
             Self::Administrator(Administrator {
                 can_manage_chat, ..
-            }) => Some(*can_manage_chat),
-            Self::Creator(_)
-            | Self::Member
-            | Self::Restricted(_)
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_manage_chat,
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => true,
         }
     }
 
-    /// Getter for [`Administrator::can_change_info`] field.
-    pub fn can_change_info(&self) -> Option<bool> {
-        match &self {
+    /// Returns `true` if the user can change the chat title, photo and other
+    /// settings.
+    ///
+    /// I.e. returns `true` if the user
+    /// - is the creator of the chat
+    /// - is an administrator in the given chat and has the [`can_change_info`]
+    ///   privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_change_info`]: Administrator::can_change_info
+    pub fn can_change_info(&self) -> bool {
+        match self {
+            Self::Creator(_) => true,
             Self::Administrator(Administrator {
                 can_change_info, ..
-            }) => Some(*can_change_info),
-            Self::Creator(_)
-            | Self::Member
-            | Self::Restricted(_)
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_change_info,
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Administrator::can_post_messages`] field.
-    pub fn can_post_messages(&self) -> Option<bool> {
-        match &self {
+    /// Returns `true` if the user can post in the channel, channels only.
+    ///
+    /// I.e. returns `true` if the user
+    /// - is the creator of the chat (even if the chat is not a channel)
+    /// - is an administrator in the given chat and has [`can_post_messages`]
+    ///   privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_post_messages`]: Administrator::can_post_messages
+    pub fn can_post_messages(&self) -> bool {
+        match self {
+            Self::Creator(_) => true,
             Self::Administrator(Administrator {
                 can_post_messages, ..
-            }) => *can_post_messages,
-            Self::Creator(_)
-            | Self::Member
-            | Self::Restricted(_)
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => can_post_messages.unwrap_or_default(),
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Administrator::can_edit_messages`] field.
-    pub fn can_edit_messages(&self) -> Option<bool> {
-        match &self {
+    /// Returns `true` if the user can edit messages of other users and can pin
+    /// messages, channels only.
+    ///
+    /// I.e. returns `true` if the user
+    /// - is the creator of the chat (even if the chat is not a channel)
+    /// - is an administrator in the given chat and has the
+    ///   [`can_edit_messages`] privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_edit_messages`]: Administrator::can_edit_messages
+    pub fn can_edit_messages(&self) -> bool {
+        match self {
+            Self::Creator(_) => true,
             Self::Administrator(Administrator {
                 can_edit_messages, ..
-            }) => *can_edit_messages,
-            Self::Creator(_)
-            | Self::Member
-            | Self::Restricted(_)
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => can_edit_messages.unwrap_or_default(),
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Administrator::can_delete_messages`] field.
-    pub fn can_delete_messages(&self) -> Option<bool> {
-        match &self {
+    /// Returns `true` if the user can delete messages of other users.
+    ///
+    /// I.e. returns `true` if the user
+    /// - is the creator of the chat
+    /// - is an administrator in the given chat and has the
+    ///   [`can_delete_messages`] privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_delete_messages`]: Administrator::can_delete_messages
+    pub fn can_delete_messages(&self) -> bool {
+        match self {
+            Self::Creator(_) => true,
             Self::Administrator(Administrator {
                 can_delete_messages,
                 ..
-            }) => Some(*can_delete_messages),
-            Self::Creator(_)
-            | Self::Member
-            | Self::Restricted(_)
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_delete_messages,
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Administrator::can_manage_voice_chats`] field.
-    pub fn can_manage_voice_chats(&self) -> Option<bool> {
-        match &self {
+    /// Returns `true` if the user can manage voice chats.
+    ///
+    /// I.e. returns `true` if the user
+    /// - is the creator of the chat
+    /// - is an administrator in the given chat and has the
+    ///   [`can_manage_voice_chats`] privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_manage_voice_chats`]: Administrator::can_manage_voice_chats
+    pub fn can_manage_voice_chats(&self) -> bool {
+        match self {
+            Self::Creator(_) => true,
             Self::Administrator(Administrator {
                 can_manage_voice_chats,
                 ..
-            }) => Some(*can_manage_voice_chats),
-            Self::Creator(_)
-            | Self::Member
-            | Self::Restricted(_)
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_manage_voice_chats,
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Administrator::can_invite_users`] field.
-    pub fn can_invite_users(&self) -> Option<bool> {
+    /// Returns `true` if the user can can invite new users to the chat.
+    ///
+    /// I.e. returns `true` if the user
+    /// - is the creator of the chat
+    /// - is an administrator in the given chat and has the [`can_invite_users`]
+    ///   privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_invite_users`]: Administrator::can_invite_users
+    pub fn can_invite_users(&self) -> bool {
         match &self {
+            Self::Creator(_) => true,
             Self::Administrator(Administrator {
                 can_invite_users, ..
-            }) => Some(*can_invite_users),
-            Self::Creator(_)
-            | Self::Member
-            | Self::Restricted(_)
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_invite_users,
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Administrator::can_restrict_members`] field.
-    pub fn can_restrict_members(&self) -> Option<bool> {
-        match &self {
+    /// Returns `true` if the user can restrict, ban or unban chat members.
+    ///
+    /// I.e. returns `true` if the user
+    /// - is the creator of the chat
+    /// - is an administrator in the given chat and has the
+    ///   [`can_restrict_members`] privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_restrict_members`]: Administrator::can_restrict_members
+    pub fn can_restrict_members(&self) -> bool {
+        match self {
+            Self::Creator(_) => true,
             Self::Administrator(Administrator {
                 can_restrict_members,
                 ..
-            }) => Some(*can_restrict_members),
-            Self::Creator(_)
-            | Self::Member
-            | Self::Restricted(_)
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_restrict_members,
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Administrator::can_pin_messages`] field.
-    pub fn can_pin_messages(&self) -> Option<bool> {
-        match &self {
+    /// Returns `true` if the user can pin messages, supergroups only.
+    ///
+    /// I.e. returns `true` if the user
+    /// - is the creator of the chat (even if the chat is not a supergroup)
+    /// - is an administrator in the given chat and has the [`can_pin_messages`]
+    ///   privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_pin_messages`]: Administrator::can_pin_messages
+    pub fn can_pin_messages(&self) -> bool {
+        match self {
+            Self::Creator(_) => true,
             Self::Administrator(Administrator {
                 can_pin_messages, ..
-            }) => *can_pin_messages,
-            Self::Creator(_)
-            | Self::Member
-            | Self::Restricted(_)
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => can_pin_messages.unwrap_or_default(),
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Administrator::can_promote_members`] field.
-    pub fn can_promote_members(&self) -> Option<bool> {
-        match &self {
+    /// Returns `true` if the user can add new administrators with a subset of
+    /// his own privileges or demote administrators that he has promoted,
+    /// directly or indirectly (promoted by administrators that were appointed
+    /// by the user).
+    ///
+    /// I.e. returns `true` if the user
+    /// - is the creator of the chat (even if the chat is not a channel)
+    /// - is an administrator in the given chat and has the
+    ///   [`can_promote_members`] privilege.
+    /// Returns `false` otherwise.
+    ///
+    /// [`can_promote_members`]: Administrator::can_promote_members
+    pub fn can_promote_members(&self) -> bool {
+        match self {
+            Self::Creator(_) => true,
             Self::Administrator(Administrator {
                 can_promote_members,
                 ..
-            }) => Some(*can_promote_members),
-            Self::Creator(_)
-            | Self::Member
-            | Self::Restricted(_)
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_promote_members,
+            Self::Member | Self::Restricted(_) | Self::Left | Self::Kicked(_) => false,
         }
     }
+}
 
-    /// Getter for [`Restricted::can_send_messages`] field.
-    pub fn can_send_messages(&self) -> Option<bool> {
+/// Methods for checking member rights.
+impl ChatMemberKind {
+    /// Returns `true` if the user can send text messages, contacts, locations
+    /// and venues.
+    ///
+    /// I.e. returns **`false`** if the user
+    /// - has left or has been kicked from the chat
+    /// - is restricted and doesn't have the [`can_send_messages`] right
+    /// Returns `true` otherwise.
+    ///
+    /// [`can_send_messages`]: Restricted::can_send_messages
+    pub fn can_send_messages(&self) -> bool {
         match &self {
             Self::Restricted(Restricted {
                 can_send_messages, ..
-            }) => Some(*can_send_messages),
-            Self::Creator(_)
-            | Self::Administrator(_)
-            | Self::Member
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_send_messages,
+            Self::Creator(_) | Self::Administrator(_) | Self::Member => true,
+            Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Restricted::can_send_media_messages`] field.
-    pub fn can_send_media_messages(&self) -> Option<bool> {
+    /// Returns `true` if the user is allowed to send audios, documents, photos,
+    /// videos, video notes and voice notes.
+    ///
+    /// I.e. returns **`false`** if the user
+    /// - has left or has been kicked from the chat
+    /// - is restricted and doesn't have the [`can_send_media_messages`] right
+    /// Returns `true` otherwise.
+    ///
+    /// [`can_send_media_messages`]: Restricted::can_send_media_messages
+    pub fn can_send_media_messages(&self) -> bool {
         match &self {
             Self::Restricted(Restricted {
                 can_send_media_messages,
                 ..
-            }) => Some(*can_send_media_messages),
-            Self::Creator(_)
-            | Self::Administrator(_)
-            | Self::Member
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_send_media_messages,
+            Self::Creator(_) | Self::Administrator(_) | Self::Member => true,
+            Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Restricted::can_send_other_messages`] field.
-    pub fn can_send_other_messages(&self) -> Option<bool> {
+    /// Returns `true` if the user is allowed to send animations, games,
+    /// stickers and use inline bots.
+    ///
+    /// I.e. returns **`false`** if the user
+    /// - has left or has been kicked from the chat
+    /// - is restricted and doesn't have the [`can_send_media_messages`] right
+    /// Returns `true` otherwise.
+    ///
+    /// [`can_send_media_messages`]: Restricted::can_send_media_messages
+    pub fn can_send_other_messages(&self) -> bool {
         match &self {
             Self::Restricted(Restricted {
                 can_send_other_messages,
                 ..
-            }) => Some(*can_send_other_messages),
-            Self::Creator(_)
-            | Self::Administrator(_)
-            | Self::Member
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_send_other_messages,
+            Self::Creator(_) | Self::Administrator(_) | Self::Member => true,
+            Self::Left | Self::Kicked(_) => false,
         }
     }
 
-    /// Getter for [`Restricted::can_add_web_page_previews`] field.
-    pub fn can_add_web_page_previews(&self) -> Option<bool> {
+    /// Returns `true` if the user is allowed to add web page previews to their
+    /// messages.
+    ///
+    /// I.e. returns **`false`** if the user
+    /// - has left or has been kicked from the chat
+    /// - is restricted and doesn't have the [`can_send_media_messages`] right
+    /// Returns `true` otherwise.
+    ///
+    /// [`can_send_media_messages`]: Restricted::can_send_media_messages
+    pub fn can_add_web_page_previews(&self) -> bool {
         match &self {
             Self::Restricted(Restricted {
                 can_add_web_page_previews,
                 ..
-            }) => Some(*can_add_web_page_previews),
-            Self::Creator(_)
-            | Self::Administrator(_)
-            | Self::Member
-            | Self::Left
-            | Self::Kicked(_) => None,
+            }) => *can_add_web_page_previews,
+            Self::Creator(_) | Self::Administrator(_) | Self::Member => true,
+            Self::Left | Self::Kicked(_) => false,
         }
     }
 }